Sanitise Bluetooth file name hint when receiving a contact
If a contact name is entered with any of the following characters

  :"<>*?|\n\t

the transfer of that contact fails.  This is due to the underlying
filesystem (FAT) not being able to handle these characters in a
file name.  This fix corrects that situation by replacing any
whitespace characters with a space and strips any illegal FAT
filesystem characters completely.

Change-Id: I5021bd26a16c31810a61bac3f70439c1153451c2