pullVcardListing(Number) when multiple contacts have same name.
When performing pullVcardListing with the search attribute
Number set to a value belonging to a contact named "A", and
there is another contact also named "A" but holding completely
different phone numbers, then the response of the
pullVcardListing procedure contains both of the "A" contacts.

Instead only the contact actually holding the Number value
that is search for should be returned, i.e. in the case above
only one contact "A".

Change-Id: Id278e1def30c8906230048fc6dd8236e94719c44