Pull vCard listing and pull vCard entry improvements
The functionality has been improved to solve the following bugs:
1 List start offset is not correct when a search value is used. The
offset is not used on the searched values but on the whole phonebook.
This is not correct according to 5.3.4.3 and 5.3.4.5 in the PBAP
specification.
2 Handle in phonebook listing is only a counter and can’t be used
to actually find the real vCard in pull vCard entry. This is clearly
visible when using a listStartOffset in the phonebook vCard listing.
This is not correct according to 3.1.5.1 in the PBAP specification.
3 Alphabetical order does not work in search by number.
The common solution for all these problems is to use the database
more extensively for lookup. Furthermore to ensure unique handles
in the phonebook listing the handle has been changed from a counter
to database ID.

Change-Id: I6cdabd98614ee602f24446e23d947f84c5b44204