bluetooth/history: moved list clearing operation to an async task
If BluetoothOppTransferHistory view contains a lenghty list of items
(downloaded/sent/failed files), clearing this list can take some time.
The problem is that the clearing is done in foreground and blocks the
application from responding, usually leading to an ANR if there's a
user sollicitation at that moment.

The fix consists in moving the clearing process to an AsyncTask.
Clearing is then done in background while application can still
treat I/O events.

Change-Id: I9a6d9c8b62a8021307906c20c4d21443d9d2536b
Author: Fabien Peix <fabienx.peix@intel.com>
Signed-off-by: Xiaokang Qin <xiaokang.qin@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 10553 17106