Holding FULL_WAKE_LOCK for 60 seconds to hold display ON
As of now BT file transfer wakes device for too short time interval
this is because of releasing FULL_WAKE_LOCK (after 1 second) and
holding the PARTIAL_WAKE_LOCK which turns off screen and keyboard.

This patch holds the FULL_WAKE_LOCK for 60 seconds so that user will
have a buffer time to look in to the display and decide weather to
accept or deny the operation.

Meanwhile since we have held the PARTIAL_WAKE_LOCK, so if user presses
power button CPU is still alive to handle the file operation.

Change-Id: I3dd4fa0badcd040a1de7422ff83b5b8fa639fafe
Acked-by: Pavan Savoy <pavan_savoy@ti.com>
Signed-off-by: halli manjunatha <hallimanju@gmail.com>