Handle message 'DISABLED' in OnState
Sometimes if bluedroid fail to disable BT in time (timeout in AdapterState,
but the operation is still in process in bluedroid), The state of AdapterState
will roll back to OnState. When the operation is done in bluedroid, it will
report a 'DISABLED' message; need to handle this message to transfer to the
right state(OffState).

Change-Id: I59001c47380c1aeae7dc3ee333191c950cd38872
Signed-off-by: yuwan@marvell.com <yuwang19860130@gmail.com>