reply with Not Found error if a phonebook object doesn't exist
  explicit weak control allows many carkits to connect to the
PBAP-PSE even if they are not strictly compliant with PBAB spec.
Moreover the error code is NOT ACCEPTABLE if the object doesn't
exist,but it should be NOT FOUND. As a result the
test TC_PSE_PBD_BI_01_C in PTS4.9 fails.
  Implemented solution: the weak control is maintained, but the
name check for phonebook object is enforced to pass
TC_PSE_PBD_BI_01_C. That test uses a fake object's name
"pbpb.vcf". In case the object doesn't exist or an invalid
phonebook name is requested, then the correct answer is
NOT FOUND. As requested by PBAB spec (6.2.3), it's mandatory for
PSE to support the "Not Found" error code.

Signed-off-by: Mirko Biselli <mirkox.biselli@intel.com>
Signed-off-by: chenxi2x <xiaolongx.chen@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Change-Id: I0d7e65cbe12855c871d86d162f5d48995b9a825d