check connection state before connecting
when two headsets are paired and connection to
the headsets is toggled rapidly, connectToOtherProfile
tries to connect device without checking the current
state of profile on one of the bonded devices.
If another device is in connecting state for the same
profile, once it gets connected, connect message arrives
and disconnects the device. This causes one device
to connect in hfp and another device to connect in a2dp.

Bug 68798

Change-Id: I568ba5e070414515497e9101750b99c9ac09e3ae