Allow BluetoothGatt to advertise all 26 bytes of man data
This addresses Issue #72304 where you cannot set more than 24 bytes of manufacturing advertising data.  The problem is that getAvailableSize() incorrectly subtracts two overhead bytes from its result if there is zero-length array in mServiceData, which is always the case if mServiceData has not been set.  The solution is to check against a zero-length array before subtracting the overhead bytes from the result.

Change-Id: I04bde32701e619c83afc056b8ea8f2414d41bf19
Signed-off-by: David G. Young <david@radiusnetworks.com>