Implement BT vendor specific HCI interface.
This change uses the api added to the BT hal layer in the
associated change in hardware/libhardware to implement the
new hidden API added in change the associated change in
frameworks/base.

This change also adds a new 'powered' state to the state machine which
represents the bluetooth controller being powere on and the stack
initialized, but no scanning or advertising modes are enabled. This is
used when the user has not enabled BT but a system app is using the
vendor-specific HCI interface.

Change-Id: I12f27c8b13fae8685412851f2b0cac2152bec330
Signed-off-by: James Bootsma <james.bootsma@thisisant.com>