Bluetooth-OPP: Fix OPP Transfer Declining
Commit e8a78ee8fa1ac47fc92f2ee8be7a50544d566b2f introduces a regression:
Unable to receive file over opp if the previous transfer has been declined.

Indeed, declining the OPP transfer just does nothing, connection is
maintained and the sender doesn't receive anything, the state becomes
inconsitent.

This patch fixes this by removing the share if tranfer is declined
(in the same way as ongoing transfer canceling).
Moreover, once file is created, we need to update the content provider
with its name, otherwise file can't be removed and a 0 Byte file remains
in sdcard/bluetooth.

Change-Id: I9276781cb989d5ac356ae2e4ac2855410eb101ef
Signed-off-by: Loic Poulain <loic.poulain@intel.com>
Signed-off-by: Yong Yao <yong.yao@intel.com>