AdvertiserManager never calls onMultiAdvertiseCallback when stopping adv.
If an application repeatedly starts and stops advertisment, a crash occurs
after few loops (less than 10 iterations).

This is because the BluetoothLeAdvertiser registers a client on each
startAdvertisment calls whereas this client is never unregistered when
stopAdvertisment is invoked. Indeed BluetoothLeAdvertiser expects the
callback onMultiAdvertiseCallback to be called with the parameters isStart
set to false when advertisment is stopped. Therefor it waits for this
callback before unregisterering the client.

Unfortunately the AdvertiserManager never calls the onMultiAdvertiseCallback
callback when advertisment is stopped. This patch fixes this.

Change-Id: I03834e36a7560bc17093506627cc509ff33c1187
Signed-off-by: Sebastien Griffoul <sebastien.griffoul@intel.com>