Virtual calls fails because SLC is not connected
Upon the connection of a BT headset during a virtual
call (e.g. Skype call), the call establishment may fail
due to a race condition with SLC connection.

This patch checks if VIRTUAL_CALL_START message is received
before SLC is established, then wait for SLC to start the
the virtual call.

Change-Id: Ica60ff8651e76471cda99ca4ef73112b340c3c63
Signed-off-by: Ludovic Cintrat <ludovicx.cintrat@intel.com>